renamed adminDashboardAction and adminDashboardReducer to allCohortsAppplications to be more descriptive and also apply to student dashboard
