+++ src/pages/AdminDashboard.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 import { connect } from "react-redux";
-import { fetchAllApps } from "../redux/actions/adminDashboardActions";
+import { fetchAllApps } from "../redux/actions/allCohortAppsActions";
 import AppsList from "../components/AppsList.js";
 
 const AdminDashboard = ({ isLoading, apps, getAllApps }) => {
