+++ src/pages/StudentDashboard.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 import { connect } from "react-redux";
-import { fetchAllApps } from "../redux/actions/adminDashboardActions";
+import { fetchAllApps } from "../redux/actions/allCohortAppsActions";
 import StudentsAppsList from "../components/StudentAppsList";
 
 const StudentDashboard = ({ isLoading, apps, getAllApps }) => {
