+++ src/redux/reducers/index.js
@@ -1,12 +1,12 @@
 import { combineReducers } from "redux";
 import datesReducer from "./datesReducer";
-import { adminDashboardReducer } from "./adminDashboardReducer";
 import adminFormReducer from "./adminFormReducer";
+import { allCohortAppsReducer } from "./allCohortAppsReducer";
 
 const rootReducer = combineReducers({
   dates: datesReducer,
   cohortInfo: adminFormReducer,
-  apps: adminDashboardReducer
+  apps: allCohortAppsReducer
 });
 
 export default rootReducer;
